refactor(app): US-055 (part) — sortir les helpers diff d'event_handlers.rs
Code-motion pur : les 5 helpers multi-worktree-diff (handle_open_multi_diff,
collect_diff_worktrees, collect_project_worktrees, collect_multiproject_groups,
open_multi_diff_for_repo) déplacés dans app/diff_view_helpers.rs. Allège
event_handlers.rs. Comportement inchangé (clippy -D warnings + tests verts).
