+++ src-app/src/app/mod.rs
@@ -11,6 +11,7 @@ pub mod constants;
 pub mod custom_buttons_modal;
 pub mod diff_sidebar;
 pub mod diff_view_actions;
+pub mod diff_view_helpers;
 pub mod drag;
 pub mod event_handlers;
 pub mod files_sidebar;
